Refactor ImportWorkbenchV2 into smaller components
Extracts the UI logic from `ImportWorkbenchV2.tsx` into four new components:
- `ImportWorkbenchHeader`: Handles header UI and scan status.
- `ImportWorkbenchSidebar`: Handles the game list sidebar.
- `ImportWorkbenchWelcome`: Handles the initial welcome/scan screen.
- `ImportWorkbenchFooter`: Handles the footer actions and progress.

This improves maintainability and readability by separating concerns and reducing the size of the main component.
